Unify object creation pattern
- For local objects (created by the interop), create the wrapper first,
then the native object which returns the native handle, then set the
native handle on the wrapper with SetHandle().
  - Example: LocalAudioTrack, AudioTransceiver created from
  AddTransceiver().

- For remote objects (created by the native implementation), use reverse
P/Invoke to create a wrapper, then create the native object, then while
firing a callback to inform the wrapper layer (C#) of the newly created
object, also pass the native handle so that the callback can call
SetHandle() on the previously created wrapper.
  - Example: RemoteAudioTrack, AudioTransceiver created as part of
  creating a RemoteAudioTrack.