Build transceiver wrappers from session description
Create the transceiver wrappers just after a local or remote session
description is applied to the peer connection, without waiting for a
track callback (remote) or taking advantage of a user call (local).

This ensures that transceiver wrappers are already created when the
sdp-ready-to-send events (both local and remote) are fired, which in
turns allows referencing those transceiver wrappers.