[Unity] Implement smart CreateOffer/Answer
Implement some smart CreateOffer and CreateAnswer methods on the
PeerConnection component, which manage transceivers on demand based on
the source components registered in the scene and their pairing name.

In Unity, the user associates to the VideoSource and AudioSource derived
classes a pairing name which the implementation automatically matches
with the equivalent components on the remote peer. For example, a
LocalVideoSource "A" is automatically paired with a RemoveVideoSource
component "A" on the remote peer.

This is achieved by passing the pairing name as the first stream ID on
the transceiver, and using this to associate the tracks signaled by the
TrackAdded event to the remote source Unity components.