[Unity] Add fake video source for demo/testing
Add a new FakeVideoSource component acting as a VideoSource and
producing uniform colored frames changing over time with a given speed.

This is mainly designed to be used as a cheap alternative to a video
capture device (webcam) for demos and tests, as it does not require any
hardware, and produces very small frames (default: 16px by 16px).