[Unity] Add hard-coded signaler for demos
Add a new HardcodedSignaler component acting as a signaling solution
between two PeerConnection components present in the same Unity scene.

This is mainly used for demos, to avoid the need to have a node-dss
server running or any other kind of networking, and make the demo
self-contained, similar to the official JavaScript WebRTC demos.