Fix race condition in external video track source
Ensure that the native track source does not start capturing before the
wrapper indicated it is ready. Otherwise the native layer may invoke
some frame request callback on the wrapper before the creation call
returned with the native interop handle for it, making it impossible for
the wrapper to complete the frame request without a native interop
handle.