Cache track name for GetName()
Cache the track name for local and remote audio and video tracks to
return as GetName() implementation, to ensure that calls to GetName()
are always safe and do not need a dispatched call to the WebRTC
signaling thread. This ensures that calling GetName() from a live object
report will always work, even with WebRTC threads destroyed. Otherwise a
call to GetName() on those objects could deadlock.