Update Unity library for split sources
Change the Unity library to build over the new split media sources. This
changes MediaLine to reference directly a media track source instead of
a media sender, and makes MediaSender and derived a slim bridge
(non-MonoBehaviour) managed by the peer connection automatically. This
way the user can still directly assign a MicrophoneSource or
WebcamSource component to a peer connection without the hassle of having
to create explicitly a track for those media track sources.